==> P0_Decompress_Data.ipynb <==
### Import Libraries for Data Preprocessing

# import libraries
import pandas as pd
import pydicom       # for reading and writing DICOM files
import numpy as np
import matplotlib.pyplot as plt
import cv2

### Read metadata 

# open metadata file
path = 'CBIS-DDSM_Data/images_dataset/metadata.csv' 
metadata_df = pd.read_csv(path)

# show content
# print(metadata_df.head(2))

# print the columns of the metadata
print(metadata_df.columns)

dirs = metadata_df["File Location"].str.replace("\\", "/", regex=False)
dirs[:3] 

dirs[0]

# prepare series and labels
series = metadata_df["Series UID"]
labels = metadata_df["Series Description"]
print(series[:3], "\n")
print(labels[:3])

print("Total number of series:", len(series))
print("Total unique series ", len(set(series)))

### Access images
- Find and build paths to download data

import os

# path to output directory
main_dir = 'CBIS-DDSM_Data/images_dataset/'

# Adds main directory parent path 
dirs_paths = [main_dir + d for d in dirs]
print("Number of directories with images", len(dirs_paths))
print(dirs_paths[0])

# Access each file in the directory 
files_per_dir = [os.listdir(path) for path in dirs_paths]
files_per_dir

# Test: try decompresing and accesing one file
# If file was accessed, no error should be printed
full_path = dirs_paths[1] + "/" + files_per_dir[0][0]
ds = pydicom.dcmread(full_path)
print(ds.file_meta.TransferSyntaxUID.name)

### Create Names for Image Files

# iterate files and series arrays
new_files_names = []

for files, serie, label in zip(files_per_dir, series, labels):
    dir_file_names = []
    
    # iterate files for each directory of series 
    if len(files) == 2: 
        # concatenate series, label and file name
        dir_file_names.append(serie + "_" + "cropped" + ".png")
        dir_file_names.append(serie + "_" + "roi" + ".png")
    elif len(files) == 1:
        # concatenate series, label and file name
        if label == "full mammogram images":
            dir_file_names.append(serie + "_" + "full" + ".png")
        elif label == "ROI mask images":
            dir_file_names.append(serie + "_" + "roi" + ".png")
        elif label == "cropped images":
            dir_file_names.append(serie + "_" + "cropped" + ".png")
            

    # append files names for directory to the array 
    new_files_names.append(dir_file_names)

count = 0
for im_dir in new_files_names:
    for file in im_dir:
        count =count + 1
print("Total number images", count)

new_files_names[:3]

### Extract Files 
Uses the data from the metadata file for extracting and decompresing images.

- Paths to images files
- The number of files that each directory contain. This help to extract ROIs and Cropped images
- The series to use as a name. However, for each series there may be more than one file. Therefore a termination to identify each image will be added.  

# Iterate through files for extracting images and saves them in directories with the  series UIDs
from pydicom.pixel_data_handlers.util import apply_voi_lut

# make images directory
output_dir = "CBIS-DDSM_Clean_Data/labeled_images_png/"
os.makedirs(output_dir, exist_ok=True)

# iterate paths, files and series arrays
for path, files, serie, names_dir in zip(dirs_paths, files_per_dir, series, new_files_names):
    # iterate files for each directory of series 
    for file, new_file_name in zip(files, names_dir):
        # concatenate path and file name
        full_path = path + "/" + file

        # read dicom file data 
        img_ds = pydicom.dcmread(full_path)

        # create pixel arrays
        pixels_arr = apply_voi_lut(img_ds.pixel_array, img_ds)

        # Normalize array
        img_pixels = (pixels_arr - pixels_arr.min()) / (pixels_arr.max() - pixels_arr.min()) * 255

        # Assigns type of array
        img_pixels = img_pixels.astype("uint8")

        # Creates file
        cv2.imwrite(os.path.join(output_dir, new_file_name), img_pixels)
