==> P2_0_Local_Preprocess.ipynb <==
# Phase 2: Find Processing Techniques
## Preprocess Images Locally
### Import Libraries and Scripts 

%reload_ext autoreload
%autoreload 2

# import functions 
from Utils.preporcessing_utils import data_loading 
from Utils.preporcessing_utils import labels_encoding
from Utils.preporcessing_utils import split_data
from Utils.preporcessing_utils import ablation
from Utils.preporcessing_utils import combination
from Utils.preporcessing_utils import preprocess_locally

### Data Preparation and Basic Preprocessing

# Loads data
train_df, test_df = data_loading("train_full.csv", "test_full.csv")

# Create dataframe and transform(encodes) pathology labels
train_df, test_df = labels_encoding(train_df, test_df)

train_df.columns

# Split data
train_data, val_data, test_data = split_data(train_df, test_df, 0.11)

### Create Preporcessing Groups By Using Ablation

# create groups using ablation
options = ['apply_background_removal',
           'apply_crop',
           'apply_noise_reduction',
           'apply_contrast_enhancement',
           'apply_edge_enhancement',
           'apply_lbp_texturizer']

# create group of techniques to try
techniques_groups = ablation(options)

techniques_groups["Baseline Basic Preporcessing"]

### Preprocesss Images Locally by using Ablation

# saves images locally 
preprocess_locally(train_data, techniques_groups)
preprocess_locally(val_data, techniques_groups)
preprocess_locally(test_data, techniques_groups)

### Create Preporcessing Groups By Using Combinations

# create groups using ablation
options_c = ['apply_background_removal',
           'apply_crop',
           'apply_noise_reduction',
           'apply_contrast_enhancement',
           'apply_edge_enhancement',
           'apply_lbp_texturizer']

# create group of techniques to try
techniques_groups = combination(options_c)

techniques_groups["Baseline Basic Preporcessing"]

### Preprocesss Images Locally by using Combination

# saves images locally 
preprocess_locally(train_data, techniques_groups)
preprocess_locally(val_data, techniques_groups)
preprocess_locally(test_data, techniques_groups)
