==> P4_Final_Comparison.ipynb <==
# Import libraries
import os
import json
import pandas as pd

# get files and directories inside Outputs
directory_path = "Outputs/"
out_files = os.listdir(directory_path)
iterations_dirs = [file for file in out_files if "Iteration" in file]
iterations_dirs

metrics_data = {}
# gets files in each iteartion directory
for name in iterations_dirs:
    dir_path = directory_path + name + "/"
    files = os.listdir(dir_path)
    # exclude checkpoints
    models_files = [file for file in files if ".ipynb_checkpoints" not in file]

    iter_data = {}
    # gets model information 
    for model_name in models_files:
        # retrieve file data 
        with open(dir_path + model_name, 'r') as f:
            model_data = json.load(f)

        # gets model name from file
        m_name = list(model_data.keys())[0]

        # gets data from file and saves it in dictionary
        data = {n: m for n, m in model_data[m_name]["metrics"].items() if n != "confusion_matrix"}
        iter_data[m_name] = data
        metrics_data[name] = iter_data

metrics_iter1 = pd.DataFrame(metrics_data["Iteration1"]).transpose()
metrics_iter1

metrics_iter2 = pd.DataFrame(metrics_data["Iteration2"]).transpose()
metrics_iter2

metrics_iter3 = pd.DataFrame(metrics_data["Iteration3"]).transpose()
metrics_iter3

metrics_iter32 = pd.DataFrame(metrics_data["Iteration3.2"]).transpose()
metrics_iter32

metrics_iter4 = pd.DataFrame(metrics_data["Iteration4"]).transpose()
metrics_iter4
